backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiogram==3.7.0
python-dotenv==1.0.1
aiofiles==23.2.1
python-multipart==0.0.9
